New hash table implementation! uthash is gone, because it used to crash in all the operations. I was unsure whether that was my fault, but now I've written a new (and very simple) one myself, nothing crashes :)
The direntry cache it it's main client, and that still needs a total re-write, tho I no longer see any crashed in it.
There is a very odd (and reproducible) crash in hash_table_del from progress, which remains unfixed.
We should rethink the use of SHA1 as the hash - it's /terribly/ slow.
uthash used to include string.h and stddef.h - these now need doing by hand in some places.
This hash table has a proper iterator!
- It needs a safe iterator too.