First step towards the direntry caching actually working. The most bogus code is gone. Straight fetching with the cache turned off works fine. Simple fetching through the cache works - if it's a hit we use it, if it's a miss we fetch it and cache it for next time. Throwing things out of the cache when they're found to no longer exist is half done - we throw the item itself out - need to start walking up the tree. Throwing things out when they may be out of date isn't done yet.
A strange problem also plagues the system when we don't have something cached (cache on or off), and the indexnode claims it exists because it hasn't updated its records yet.
Here, we make a new de in fetch_node, based on what the indexnode tells us, and immediately try to throw it away when it doesn't exist.
Creation posts it (this is the one read() gets, if you like), and the downloader posts it. The downloader deletes it when it doesn't work, and so does read when it quits.
If the cache is on, there are another pair from cache_add and cache_del (called by notify stale).
These seem to tally, but glibc complains of a double free.