Big, but actually quite small (actually quite big) change to direntry caching.
The upshot is that the cache now links everything it adds to its parent, thus constructing a tree structure over the items in the cache. This is needed before we can take the next step in the cache re-write, namely being able properly to remove stale parts of the tree.

This has required a few changes:
- We must now fetch all children of a node if we fetch any, otherwise you get into the position of trying to add duplicates to the cache, which would not be nice to deal with.
- Nothing but the cache can now add parent <-> child relationships otherwise things would be awkward.
- Sibling relationships (on which this tree structure also depends) are done outside of the cache but are guaranteed to be sufficient by the fact that we fetch everything at once.
- Enumerating the children of a directory now returns a list of them, not a parent with them attached (they may now not be attached to their parent if you got hold of it).

Also, added more hash table and direntry cache tree debugging functionality.