Direntry cache is now informed when an attempt to fetch something it's got cached (either read a file it knows about, or list a directory it knows about as a leaf).
In response, it pokes the indexnode to see what in that region of the tree is and isn't actually in existence any more and pu     rges old stuff.

direntry cache now now has a trace area.

flush() and release() calls - interesting to see how the correspond to our direntry_post()/delete()ing.

cache_notify_stale() should lock the direntry cache, as it's going to iterate it, but this would require locks to be recursive as that function manipulated the cache.