Statically configures indexnodes are now ping'd.
Rather than just instantiating them at startup, they are added to a list and pinged (more frequently than the timeout).
The idea is that they'd be statically configured if we can't see their broadcasts, but they (or our route to them) may still come and go.
Thus /we/ go to /them/ before they're due to expire and see if they're still reachable.
If they are, we raise the event to say they've been seen, just as we do in response to an advert broadcast.